Update ecommerce.html
here on the e-commerce site, the admin can keep track of the products searched and there is a button called add product. when a user clicks on add product, the search entry gets stored below the form with a delete button and an edit button. when a user clicks on the delete, the product added by them is deleted and to modify the product entry the user can click on the edit button the entry gets deleted and the amount and product tab are populated with the user's previous entry and there we can edit their last entry. the total amount gets updated and sums up to the amount entered by various users. and once a user clicks on the delete button or tries to modify their search by clicking the edit button, that particular user's amount entered gets subtracted from the total amount and the updated value of the total amount is shown on the screen.